requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
beautifulsoup4==4.12.2
nltk==3.8.1
numpy==1.24.3
pandas==2.0.1
Requests==2.30.0
